Last commit before rewrite
This is the last commit for a while.

I will be creating a new branch `rewrite` and working on it from now on.
The objective is to completely overhaul the Scene system and replace it with a GameState mechanism. Hopefully, this won't take long. Each GameState will have extended features over a scene, including persistence, layering, instance-level UI managers, etc